create controller repositort service para a classe animal

+++ services/proprietarioService.js
@@ -1,4 +1,6 @@
 import ProprietarioRepository from '../repositories/proprietarioRepository.js';
+import AnimalRepository from '../repositories/animalRepository.js';
+import animalRepository from '../repositories/animalRepository.js';
 
 async function createProprietario(proprietario) {
     return await ProprietarioRepository.insertProprietario(proprietario);
@@ -17,6 +19,13 @@ async function getProprietario(id) {
 }
 
 async function deleteProprietario(id) {
+    const animais = await AnimalRepository.getAnimalByProprietario(id);
+    if (animais.length > 0) {
+        throw new Error(
+            'Este proprietario não pode ser excluido por ter algum pet aqui.'
+        );
+    }
+    console.log(animais);
     await ProprietarioRepository.deleteProprietario(id);
 }
